Split cipher_light.h into public+internal
We'll need the type definition in gcm.h which is public.
